feat: merge migration files and add attached_image_filename to resource_comment_reply and utilization_comment_reply
